contrib/pagestore: extract the shared daemon brain into pagestore_core (S1.2a)
Toward two daemon binaries that share one brain: a portable POSIX daemon and an
optional SPDK daemon.  Move everything backend- and loop-agnostic into
pagestore_core.{c,h}: the in-memory page/fork indexes, copy-on-write
read_through, timelines, the per-page WAL index, shipped-WAL metadata, recovery,
and ps_handle_meta() which handles every request that is not page byte I/O.

pagestore_daemon.c becomes a thin POSIX frontend: the synchronous shared-memory
channel loop plus the four byte-I/O ops (EXTEND/WRITEV/READV/READ_AT) served via
append_page/read_through/read_version against the POSIX storage backend.  The
seam is ps_handle_meta + the read/append helpers, so the SPDK daemon can reuse
pagestore_core.c verbatim with an asynchronous loop and the SPDK storage backend.

meson and the CI cc line build pagestore_core.c alongside the daemon.  Pure
refactor, no behaviour change: standalone suite 110/110, integration_test PASS.
